+++ collections/settings.js
@@ -111,7 +111,7 @@ settingsSchemaObject = {
     autoform: {
       group: 'posts',
       instructions: 'The view used for the front page',
-      options: _.map(viewNav, function (view) {
+      options: _.map(viewsMenu, function (view) {
         return {
           value: camelCaseify(view.label),
           label: view.label