+++ packages/telescope-base/package.js
@@ -15,8 +15,8 @@ Package.onUse(function (api) {
     'STATUS_APPROVED',
     'STATUS_REJECTED',
     
-    'adminNav', 
-    'viewNav', 
+    'adminMenu', 
+    'viewsMenu', 
     'addToPostSchema', 
     'addToCommentsSchema', 
     'addToSettingsSchema',