+++ packages/telescope-daily/lib/daily.js
@@ -1,4 +1,4 @@
-viewNav.push({
+viewsMenu.push({
   route: 'postsDaily',
   label: 'daily',
   description: 'day_by_day_view'