+++ packages/telescope-search/lib/search.js
@@ -4,7 +4,7 @@ primaryNav.push({
   order: 100
 });
 
-adminNav.push({
+adminMenu.push({
   route: 'searchLogs',
   label: 'search_logs',
   description: 'see_what_people_are_searching_for'