+++ History.md
@@ -9,6 +9,9 @@
 * Added new Events collection for keeping track of important events.
 * Log first run event.
 * `MAIL_URL` environment variable can now be set from Settings.
+* Renamed `viewNav` to `viewsMenu`.
+* Renamed `adminNav` to `adminMenu`.
+* **New and improved [documentation](http://telesc.pe/docs).**
 
 ## v0.11.1 “FeedScope”